refactor(inventory): 발주 쓰기를 PurchaseOrderManager 로, 서비스를 포트로 (#724 항목 5-b)
warehouse-transfer 3계층 선례를 따른다 — service 는 위임만, manager 가 검증·쓰기,
reader 가 조회. 메서드 본문 변경 0줄.

🔴 잠금 불변식(PO 행 → 라인 행)을 세 파일 docstring 에 명시적으로 옮겼다. Manager 는
"이 파일에 쓰기를 추가하면 PO 행부터 잡는다", Reader 는 ".for('update') 를 넣는 순간
적용 대상이니 Manager 로 옮겨라", Service 는 "DB 접근을 추가할 거면 Manager 로 가라".
취득 지점 3곳에 테스트가 없고 주석만이 방어선이라 분할로 흩어지면 사라질 규약이었다.

단위 스펙을 포트가 아니라 Manager/Reader 로 겨눈다(purchase-order.manager.spec.ts).
포트는 위임만 하므로 거기서 테스트하면 배선을 두 번 재는 셈이다. 통합 스펙 2개는
buildPurchaseOrderPort 헬퍼로 DI 가 하는 조립을 손으로 한다.

971줄 God service → service 102 · manager 515 · reader 204 · cart 265 · reorder 75.

검증: type-check 0 · admin-web tsc 0 · eslint 0 · jest 489 suite · nest build 성공
· 통합 8 suite/12 test(기준선과 동일)이며 발주 통합 스펙 2개 PASS
