+++ apps/core/src/modules/inventory/procurement/services/purchase-order-line-execution.integration.spec.ts
@@ -7,10 +7,23 @@ import { wmsSchema, wmsTables, DbTx } from '../../schema/inventory.schema';
 import { makeDb, inRollbackTx } from '../../../fulfillment/services/__support__';
 import { PurchaseOrderService } from './purchase-order.service';
 import { PurchaseOrderReader } from './purchase-order.reader';
+import { PurchaseOrderManager } from './purchase-order.manager';
 import { InboundService } from '../../inbound/services/inbound.service';
 import { PurchaseOrderStatus, PurchaseOrderType } from '../dto/purchase-order.dto';
 import { InboundPipelineReader } from '../../stock-projection/services/inbound-pipeline.reader';
 import { WarehouseTransferReader } from '../../warehouse-transfer/services/warehouse-transfer.reader';
+
+/**
+ * 포트 조립. 3계층(#724 항목 5-b) 이후 `PurchaseOrderService` 는 위임만 하므로 통합
+ * 스펙도 Manager/Reader 를 손으로 엮어 넣는다 — 실제 DI 가 하는 일과 같다.
+ */
+function buildPurchaseOrderPort(
+  dbService: ConstructorParameters<typeof PurchaseOrderManager>[0],
+  inboundService: ConstructorParameters<typeof PurchaseOrderManager>[1],
+): PurchaseOrderService {
+  const reader = new PurchaseOrderReader(dbService);
+  return new PurchaseOrderService(new PurchaseOrderManager(dbService, inboundService, reader), reader);
+}
 
 /**
  * 발주 라인을 하나씩 실제 발주 실행하는 경로를 고정한다.
@@ -65,7 +78,7 @@ describeIfDb('발주 라인 실행 (DB integration)', () => {
 
   function buildService(trx: DbTx): PurchaseOrderService {
     const dbService = boundDbService(trx);
-    return new PurchaseOrderService(dbService, buildInboundService(trx), new PurchaseOrderReader(dbService));
+    return buildPurchaseOrderPort(dbService, buildInboundService(trx));
   }
 
   interface Fixture {
